rename firebase file
