+++ lib/firebase/firebase.ts
@@ -1,5 +1,6 @@
 import { cert, getApps, initializeApp } from "firebase-admin/app";
 import { getFirestore } from "firebase-admin/firestore";
+import { getMessaging } from "firebase-admin/messaging";
 
 const useFirebaseAdmin = () => {
   const serviceAccount: string = process.env.GOOGLE_APPLICATION_CREDENTIALS!;
@@ -12,10 +13,10 @@ const useFirebaseAdmin = () => {
         });
 
   const db = getFirestore(fireStoreApp);
-
+  const messaging = getMessaging(fireStoreApp);
   return {
-    firebaseApp: fireStoreApp,
     db,
+    messaging,
   };
 };
